fix remaining test issues
