+++ src/encoded/tests/test_lucene_builder.py
@@ -36,10 +36,11 @@ VARIANT_SAMPLE_FACETS = ['embedded.AF', 'embedded.DP', 'embedded.FS', 'embedded.
                          'embedded.variant.CHROM.raw', 'embedded.variant.POS', 'embedded.variant.cadd_phred',
                          'embedded.variant.clinvar_clnsig.raw', 'embedded.variant.conservation_phylop100',
                          'embedded.variant.cytoband_cytoband.raw',
-                         'embedded.variant.genes.genes_ensg.gene_lists.display_title.raw',
-                         'embedded.variant.genes.genes_ensg.oe_lof',
-                         'embedded.variant.genes.genes_ensg.oe_lof_upper', 'embedded.variant.genes.genes_ensg.oe_mis',
-                         'embedded.variant.genes.genes_ensg.s_het',
+                         'embedded.variant.genes.genes_most_severe_gene.gene_lists.display_title.raw',
+                         'embedded.variant.genes.genes_most_severe_gene.oe_lof',
+                         'embedded.variant.genes.genes_most_severe_gene.oe_lof_upper',
+                         'embedded.variant.genes.genes_most_severe_gene.oe_mis',
+                         'embedded.variant.genes.genes_most_severe_gene.s_het',
                          'embedded.variant.genes.genes_most_severe_consequence.coding_effect.raw',
                          'embedded.variant.genes.genes_most_severe_consequence.impact.raw',
                          'embedded.variant.genes.genes_most_severe_consequence.location.raw',
